# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Hello Books",
        "description": "A system that allows efficient, cost-effective management of our library.",
        "version": "1.0.0"
    },
    "basePath": "/api/v1/",
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/authors": {
            "post": {
                "tags": ["authors"],
                "summary": "Add a new author",
                "operationId": "addAuthor",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User object that needs to be added to the app",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewAuthor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Author added successfully"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "get": {
                "tags": ["authors"],
                "summary": "Get all authors",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "page",
                        "type": "integer",
                        "description": "Define the page index for the records",
                        "in": "query"
                    },
                    {
                        "name": "limit",
                        "type": "integer",
                        "description": "Define the limit for the record",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully fetched all authorss"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "NewAuthor": {
            "type": "object",
            "required": ["name"],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Dangote Jnr"
                }
            },
            "xml": {
                "name": "NewAuthor"
            }
        },
...
    }
}